Current State
Current state has a brand new issue I don't know why or how but it is
reconnecting each time the page is loaded, and so, it is trying to
restart and run constantly. Consider an icon on the screen that has
status of connected. When connected, the user can click the button and
it will start the game. (player 1 will have this ability)